[fix] add AITaskFailedEvent class and update event handling in AICompensationListener and RedisStreamConsumer
